+++ backend/queue/queue_test.go
@@ -2,7 +2,6 @@ package queue
 
 import (
 	"context"
-	"fmt"
 	"sync"
 	"testing"
 	"time"
@@ -20,8 +19,8 @@ func TestEnqueue(t *testing.T) {
 	client, err := e.NewClient()
 	require.NoError(t, err)
 
-	queue := New("testenq", client)
-	err = queue.Enqueue(context.Background(), "test value")
+	queue := New("testenq", client, time.Second)
+	err = queue.Enqueue(context.Background(), "test item")
 	assert.NoError(t, err)
 }
 
@@ -33,17 +32,17 @@ func TestDequeueSingleItem(t *testing.T) {
 	client, err := e.NewClient()
 	require.NoError(t, err)
 
-	queue := New("testdeq", client)
-	err = queue.Enqueue(context.Background(), "test value")
+	queue := New("testdeq", client, time.Second)
+	err = queue.Enqueue(context.Background(), "test single item dequeue")
 	require.NoError(t, err)
 
-	value, err := queue.Dequeue(context.Background())
+	item, err := queue.Dequeue(context.Background())
 	require.NoError(t, err)
-	require.Equal(t, "test value", value.Value())
+	require.Equal(t, "test single item dequeue", item.Value)
 
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
 	defer cancel()
-	_, err = queue.Dequeue(ctx)
+	item, err = queue.Dequeue(ctx)
 	require.Error(t, err)
 }
 
@@ -55,7 +54,7 @@ func TestDequeueFIFO(t *testing.T) {
 	client, err := e.NewClient()
 	require.NoError(t, err)
 
-	queue := New("testfifo", client)
+	queue := New("testfifo", client, time.Second)
 	items := []string{"hello", "there", "world", "asdf", "fjdksl", "lalalal"}
 
 	for _, item := range items {
@@ -65,9 +64,9 @@ func TestDequeueFIFO(t *testing.T) {
 	result := []string{}
 
 	for range items {
-		value, err := queue.Dequeue(context.Background())
+		item, err := queue.Dequeue(context.Background())
 		require.NoError(t, err)
-		result = append(result, value.Value())
+		result = append(result, item.Value)
 	}
 
 	require.Equal(t, items, result)
@@ -81,7 +80,7 @@ func TestDequeueParallel(t *testing.T) {
 	client, err := e.NewClient()
 	require.NoError(t, err)
 
-	queue := New("testparallel", client)
+	queue := New("testparallel", client, time.Second)
 	items := map[string]struct{}{
 		"hello":   struct{}{},
 		"there":   struct{}{},
@@ -109,11 +108,10 @@ func TestDequeueParallel(t *testing.T) {
 
 	for range items {
 		go func() {
-			value, err := queue.Dequeue(context.Background())
+			item, err := queue.Dequeue(context.Background())
 			require.NoError(t, err)
-			fmt.Println(value)
 			mu.Lock()
-			results[value.Value()] = struct{}{}
+			results[item.Value] = struct{}{}
 			mu.Unlock()
 			wg.Done()
 		}()
@@ -132,20 +130,20 @@ func TestNack(t *testing.T) {
 	client, err := e.NewClient()
 	require.NoError(t, err)
 
-	queue := New("testnack", client)
-	err = queue.Enqueue(context.Background(), "test value")
+	queue := New("testnack", client, time.Second)
+	err = queue.Enqueue(context.Background(), "test item")
 	require.NoError(t, err)
 
-	value, err := queue.Dequeue(context.Background())
+	item, err := queue.Dequeue(context.Background())
 	require.NoError(t, err)
-	require.Equal(t, "test value", value.Value())
+	require.Equal(t, "test item", item.Value)
 
-	err = value.Nack(context.Background())
+	err = item.Nack(context.Background())
 	require.NoError(t, err)
 
-	value, err = queue.Dequeue(context.Background())
+	item, err = queue.Dequeue(context.Background())
 	require.NoError(t, err)
-	require.Equal(t, "test value", value.Value())
+	require.Equal(t, "test item", item.Value)
 }
 
 func TestAck(t *testing.T) {
@@ -156,19 +154,68 @@ func TestAck(t *testing.T) {
 	client, err := e.NewClient()
 	require.NoError(t, err)
 
-	queue := New("testack", client)
-	err = queue.Enqueue(context.Background(), "test value")
+	queue := New("testack", client, time.Second)
+	err = queue.Enqueue(context.Background(), "test item")
 	require.NoError(t, err)
 
-	value, err := queue.Dequeue(context.Background())
+	item, err := queue.Dequeue(context.Background())
 	require.NoError(t, err)
-	require.Equal(t, "test value", value.Value())
+	require.Equal(t, "test item", item.Value)
 
-	err = value.Ack(context.Background())
+	err = item.Ack(context.Background())
 	require.NoError(t, err)
 
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
 	defer cancel()
-	value, err = queue.Dequeue(ctx)
+	item, err = queue.Dequeue(ctx)
 	require.Error(t, err)
+}
+
+func TestOnce(t *testing.T) {
+	t.Parallel()
+
+	e, cleanup := etcd.NewTestEtcd(t)
+	defer cleanup()
+	client, err := e.NewClient()
+	require.NoError(t, err)
+
+	queue := New("testonce", client, time.Second)
+
+	err = queue.Enqueue(context.Background(), "test item")
+	require.NoError(t, err)
+
+	item, err := queue.Dequeue(context.Background())
+	require.NoError(t, err)
+	require.Equal(t, "test item", item.Value)
+
+	// nack should return the original itemt to the queue for reprocessing, ack
+	// called after should have no effect
+	require.NoError(t, item.Nack(context.Background()))
+	require.NoError(t, item.Ack(context.Background()))
+	nackedItem, err := queue.Dequeue(context.Background())
+	require.NoError(t, err)
+	require.Equal(t, item.Value, nackedItem.Value)
+}
+
+func TestNackExpired(t *testing.T) {
+	t.Parallel()
+	e, cleanup := etcd.NewTestEtcd(t)
+	defer cleanup()
+	client, err := e.NewClient()
+	require.NoError(t, err)
+
+	timeout := 5 * time.Millisecond
+	queue := New("testexpired", client, timeout)
+
+	err = queue.Enqueue(context.Background(), "test item")
+	require.NoError(t, err)
+
+	_, err = queue.Dequeue(context.Background())
+	require.NoError(t, err)
+
+	// nacked item should be back in the queue
+	item, err := queue.Dequeue(context.Background())
+	require.NoError(t, err)
+
+	require.Equal(t, "test item", item.Value)
 }